requirements.txt
langchain
langchain-ollama
python-dotenv
langchain_community
streamlit